pyproject.toml
[project]
name = "skku-llm-ddp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0" 
dependencies = [
    "django (==5.1.6)",
    "djangorestframework (==3.15.2)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "python-dotenv (==1.0.1)",
    "mysqlclient (==2.2.7)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "faker (>=37.1.0,<38.0.0)",
    "langchain (==0.3.23)",
    "openai (>=1.70.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langgraph (>=0.3.25,<0.4.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
